Refactor selected_prices method to use ActiveRecord query
To improve the way price IDs are processed. Instead of splitting a string of
price IDs, the method now directly maps the incoming `price_ids` parameter to
integers and retrieves the corresponding prices using an ActiveRecord query.

As the user could select multiple prices on sale price tabs and the parameter was receiving
multiple price IDs as well but when the sale price was created it was created for the first
ID itself hence to align the multiple selected prices to have a sale price created by user the
query is updated.